Further optimize SHORTEST-PATHS by studying level by level first
Now we first study what certificate path level a key is found first.
This information helps in recursive route finding task because we can
stop if a key can be found on earlier level somewhere else.

This change probably does not increase the total speed of the function
in practical sense (it's already "instant"). The purpose is to reduce
the amount of duplicate work made by recursive path search. This is
friendlier for the stack used by recursive function calls.